backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
websockets = "^13.1"
requests = "^2.32.3"
tqdm = "^4.66.5"
aiohttp = "^3.10.9"
apscheduler = "^3.10.4"
aiohttp-cors = "^0.7.0"
vonage = "^3.17.1"
pystac-client = "^0.8.3"
planetary-computer = "^1.0.0"
odc-stac = "^0.3.10"
matplotlib = "^3.9.2"
ogr = "^0.50.0"
shapely = "^2.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
